refactor(electron): extract ChartService and DataTableService to separate business logic from IPC handlers
- Moved chart-related logic from ChartAPIHandlers to new ChartService
- Moved data table logic from DataTableAPIHandlers to new DataTableService
- Simplified IPC handlers to delegate to service layer
- Improved separation of concerns and maintainability

---

refactor(electron)：抽離 ChartService 與 DataTableService，分離業務邏輯與 IPC handlers

- 將圖表相關邏輯從 ChartAPIHandlers 移至新的 ChartService
- 將資料表相關邏輯從 DataTableAPIHandlers 移至新的 DataTableService
- 簡化 IPC handlers，僅負責調用 service 層
- 提升程式結構清晰度與可維護性
